replace legacy AbstractNode with GSOC version. In other words, NewAbstractNode renamed to AbstractNode, legacy AbstractNode deleted
