+++ engine/src/main/java/org/terasology/rendering/dag/stateChanges/SetInputTextureFromFbo.java
@@ -30,7 +30,7 @@ import static org.lwjgl.opengl.GL11.GL_TEXTURE_2D;
 import static org.lwjgl.opengl.GL11.glBindTexture;
 import static org.lwjgl.opengl.GL13.GL_TEXTURE0;
 import static org.lwjgl.opengl.GL13.glActiveTexture;
-import static org.terasology.rendering.dag.AbstractNode.getMaterial;
+import static org.terasology.rendering.dag.gsoc.AbstractNode.getMaterial;
 import static org.terasology.rendering.opengl.fbms.DisplayResolutionDependentFbo.POST_FBO_REGENERATION;
 
 // TODO: split this class into two - one for opengl's global state change and one for the specific material state change.
